Feature: Added a bunch of new methods, a new type, a small helper trait for toStringArray and a separate exception for DuplicateValue Feature: Added the following methods to all the array types (`IsArrayEnumType`, `IsCollectionType`, `IsIntArrayEnumType`, `IsStringArrayEnumType`, `IsClassCollectionType`): - `count`: Returns the number of elements inside the array value object. - `isEmpty`: Whether the value object contains any elements. - `isNotEmpty`: The opposite of `isEmpty`. - `isEqualTo`: Whether the value object's elements are equal to the argument. The argument can be another class (with a `toArray` method), an array, or an object with public properties. - `isNotEqualTo`: The opposite of `isEqualTo`. - `empty`: A factory method to create a new instance with no elements.
Feature: Added a new value type: `IsClassCollectionType`.
This adds on to `IsCollectionType` and makes it easier to create a collection holding an array of objects. It validates whether the items are of a specific class. By default, it does not allow duplicate values, however this can be overridden.

Dev: Separate exception for duplicate values: `DuplicateValue`.
It is backwards compatible as it extends from `InvalidValue`, but it does now allow developers to catch this exception separately.