Tests: Upgraded from Phpunit 10.5 to 11.5 (this was a forced update, as the testdox option was removed in 10.5 but brought back in 11.5)* Tests: Run Infection and fixed any true issues (ignoring false positives of course) Tests: Changed from using @covers, @uses, @dataProvider, @depends to using relevant attributes
Running the tests with testdox means that only errors are printed, not every single assertion with a tick or an X.
For over 1,500 assertions, it meant that finding errors was incredibly frustrating and time-consuming.
In PhpUnit 11.5 it is called testdoxSummary, as opposed to testdox (as it was called in PHPUnit 8), and isn't available at all in PhpUnit 9.5.